Added complete code but facing some errors need to fix

+++ src/main/java/com/batch/repo/PersonRepo.java
@@ -3,7 +3,7 @@ package com.batch.repo;
 import com.batch.entity.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonRepo extends JpaRepository<Person, Integer> {
+public interface PersonRepo extends JpaRepository<Person, Long> {
